requirements.txt
Pillow==5.4.1
tensorflow==1.12.0
pandas==0.23.4
numpy==1.15.4
matplotlib==2.2.3
pytesseract==0.2.6
opencv-python==4.0.0.21
pylint==2.2.2
pylint==2.2.2
